requirements.txt
tensorflow==2.16.1
pandas==1.5.3
matplotlib
sckit_learn==1.2.2
scipy==1.13.0
numpy==1.24.3
imbalanced-learn==0.12.0
seaborn==-0.13.2
opencv-python==4.9.0.80
pillow==10.2.0
ultralytics==8.2.2
gensim==4.3.2
python-Levenshtein==0.25.1
tensorflow-model-optimization
tensorflow-addons
notebook